+++ CHANGELOG.md
@@ -3,7 +3,16 @@
 # CodeXP Changelog
 
 ## [Unreleased]
+
 ### Added
+
+- Add CodeXP configuration in IDE setting
+- Add checkbox for show/hide completed challenges
+
+## [1.0.0] - 2023-06-04
+
+### Added
+
 - Implement core logic of CodeXP
 - Implement CodeXP Dashboard
 - Define events to detect and add challenges for each event